+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.3]
 ### Fixed
-- Changed Solarium client to use post instead of get to stop URI Too Long errors.
+- Changed Solarium client to use post instead of get to stop URI Too Long errors (PR #10)
 
 ## [1.0.2] - 2020-05-07
 ### Fixed