+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated the plugin link to point to GitHub instead of GitLab (PR #11)
 
 ## [1.0.3]
 ### Fixed
-- Changed Solarium client to use post instead of get to stop URI Too Long errors.
+- Changed Solarium client to use post instead of get to stop URI Too Long errors (PR #10)
 
 ## [1.0.2] - 2020-05-07
 ### Fixed