+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix re-indexing of blogs and update solarium version (PR #14)
+
 ### Changed
 - Updated the plugin link to point to GitHub instead of GitLab (PR #11)